refactor: move n_files function to a separate count_n_files.py in utils directory
